Fixed alignment in about page.
I fixed your images alignment, but in doing so, I went with a differnet
trasition than yours.  Your has so attributes that were conflicting with
your images, so it was just easier for me to write a little something
out then trying to find and fix that bug.  The same amount of coding
would of be done, but the trasitions are different.